refactor: move components and styles into colocated files
